Added button press functionality on admin inventory button
When pressing the inventoy button on the admin page, the app will switch to the inventory screen
