--- Controller/adminMainScreenController.py
@@ -1,6 +1,0 @@
-import kivy
-from kivy.uix.screenmanager import Screen
-
-class Controller():
-    def switch_screen(self, screen):
-        pass
